@@ -13,12 +13,12 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^8.2.7",
-    "@angular/core": "^8.2.8",
-    "@angular/forms": "^8.2.7",
-    "@angular/platform-browser": "^8.2.7",
-    "@angular/platform-browser-dynamic": "^8.2.7",
-    "@angular/router": "^8.2.7",
+    "@angular/common": "^9.0.7",
+    "@angular/core": "^9.0.7",
+    "@angular/forms": "^9.0.7",
+    "@angular/platform-browser": "^9.0.7",
+    "@angular/platform-browser-dynamic": "^9.0.7",
+    "@angular/router": "^9.0.7",
     "@capacitor/android": "^1.5.0",
     "@capacitor/core": "1.5.0",
     "@capacitor/ios": "^1.5.0",
@@ -38,9 +38,9 @@
     "@angular-devkit/core": "~9.0.7",
     "@angular-devkit/schematics": "~9.0.7",
     "@angular/cli": "~9.0.7",
-    "@angular/compiler": "~8.2.14",
+    "@angular/compiler": "~9.0.7",
     "@angular/compiler-cli": "^9.0.6",
-    "@angular/language-service": "~8.2.14",
+    "@angular/language-service": "~9.0.7",
     "@capacitor/cli": "1.5.0",
     "@ionic/angular-toolkit": "^2.1.2",
     "@types/jasmine": "~3.3.8",
